refactor(back/routes): put nextcloud-related routes into another file
