+++ back/app/routes/user.py
@@ -119,86 +119,3 @@ async def delete_reminder():
         return JSONResponse(content={"Response":"OK"})
     except Exception as e:
         raise HTTPException(status_code=500, detail=str(e))
-
-# CLOUD
-@router.get("/cloud/", description="send an address toward NextCloud's file system")
-async def get_ns_cloud():
-    try:
-        return JSONResponse(content={"Response":"OK"})
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
-
-@router.get("/cloud/?dir={dir}", description="send an address toward NextCloud's file system")
-async def get_ns_specific_folder(dir):
-    try:
-        return JSONResponse(content={"Response":"OK"})
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
-
-@router.get("/cloud/?dir={dir}&file={file}", description="send an address towards specific file into Nextcloud")
-async def get_ns__specific_file(dir, file):
-    try:
-        return JSONResponse(content={"Response":"OK"})
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
-
-@router.get("/cloud/search/{filename}", description="send an address towards Nextcloud's search file system")
-async def get_ns__specific_file(dir, file):
-    try:
-        return JSONResponse(content={"Response":"OK"})
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-@router.get("/file/{id}", description="send an address toward Nextcloud's file viewer")
-async def get_ns_file_viewer(id):
-    try:
-        return JSONResponse(content={"Response":"OK"})
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-
-@router.get("/calendar/", description="send an address toward NextCloud's calendar system")
-async def get_ns_calendar():
-    try:
-        return JSONResponse(content={"Response":"OK"})
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
-
-@router.get("/calendar/view={viewType}", description="send an address toward NextCloud's calendar system with specific view type")
-async def get_ns_specific_calendar(viewType):
-    try:
-        return JSONResponse(content={"Response":"OK"})
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-@router.get("/contact/", description="send an address toward NextCloud's contact system")
-async def get_ns_contacts():
-    try:
-        return JSONResponse(content={"Response":"OK"})
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
-
-@router.get("/contact/{contact_id}", description="send an address toward NextCloud's contact system with a specific target")
-async def get_ns_specific_contacts(contact_id):
-    try:
-        return JSONResponse(content={"Response":"OK"})
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
-
-
-@router.get("/notes/", description="send an address toward NextCloud's notes system")
-async def get_ns_notes():
-    try:
-        return JSONResponse(content={"Response":"OK"})
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
-
-@router.get("/notes/{note_id}", description="send an address toward NextCloud's notes system with a specific target")
-async def get_ns_specific_note(note_id):
-    try:
-        return JSONResponse(content={"Response":"OK"})
-    except Exception as e:
-        raise HTTPException(status_code=500, detail=str(e))
